# lib/api-zod/src/generated/types/consentDocument.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentDocument {
  id: string;
  videoId: string;
  creatorId?: string | null;
  fileUrl: string;
  fileName: string;
  contentType: string;
  uploadedBy?: string | null;
  createdAt: Date;
}