# lib/api-zod/src/generated/types/consentDocumentRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentDocumentRequest {
  fileUrl: string;
  fileName: string;
  contentType: string;
  creatorId?: string | null;
}