# lib/api-zod/src/generated/types/creatorEarnings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */
import type { EarningsByVideo } from './earningsByVideo';

export interface CreatorEarnings {
  creatorId: string;
  handle: string;
  totalEarnedCents: number;
  paidOutCents: number;
  availableCents: number;
  byVideo: EarningsByVideo[];
}