# lib/api-zod/src/generated/types/creatorRanking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatorRanking {
  rank: number;
  creatorId: string;
  handle: string;
  displayName: string;
  avatarUrl: string;
  revenueCents: number;
  purchases: number;
}