# lib/api-zod/src/generated/types/payout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface Payout {
  id: string;
  creatorId: string;
  amountCents: number;
  status: string;
  provider?: string | null;
  providerRef?: string | null;
  periodStart?: Date | null;
  periodEnd?: Date | null;
  createdAt: Date;
}