# lib/api-zod/src/generated/types/payoutCreated.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface PayoutCreated {
  id: string;
  creatorId: string;
  amountCents: number;
  status: string;
  statusLabel: string;
  createdAt: Date;
}