# lib/api-zod/src/generated/types/payoutRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface PayoutRequest {
  /** Optional. Defaults to the full available balance. */
  amountCents?: number;
}