# lib/api-zod/src/generated/types/videoStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bimboy Studios API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoStats {
  videoId: string;
  clicks: number;
  views: number;
  purchases: number;
  grossRevenueCents: number;
  /** purchases / clicks (0-1) */
  conversionRate: number;
}